+++ client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "torque-task-client",
-  "version": "1.0.5",
+  "version": "1.0.7",
   "private": true,
   "author": "TexasTorque",
   "license": "MIT",
