+++ client/src/index.css
@@ -47,6 +47,7 @@ h1,
 h2,
 h3,
 h4,
+h5, h6,
 p,
 label {
   color: white !important;
